serverless/requirements.txt
requests==2.25.1
selenium==3.141.0
GitPython==3.1.18